Some slight reorganization
