+++ frontend/src/components/homepage/Homepage.tsx
@@ -1,51 +1,28 @@
-import React, { memo, useEffect, useState } from 'react';
+import React, { memo, useEffect, useState, useCallback } from 'react';
 import styles from './styles/Homepage.module.scss';
-import Settings from '../shared/Settings';
-import SockJsClient from 'react-stomp';
+import { useHistory } from 'react-router-dom';
 
 export const Homepage = memo(function Homepage() {
-    const [clientConnected, setClientConnected] = useState(false);
-    const [messages, setMessages] = useState<String[]>([]);
-    const [clientRef, setClientRef] = useState<any | undefined>(undefined);
+    const history = useHistory();
 
-    const onMessageRecieve = (msg) => {
-        setMessages((prevMessages) => [...prevMessages, msg]);
-    };
-    const sendMessage = () => {
-        console.log('send hello');
-        try {
-            clientRef.sendMessage('/app/hello');
-            return true;
-        } catch (e) {
-            return false;
-        }
-    };
+    const navigateToCompetitiveSettings = useCallback(
+        () => history.push('/competitive-settings'),
+        []
+    );
+    const navigateToCollaborativeSettings = useCallback(
+        () => history.push('/collaborative-settings'),
+        []
+    );
     return (
         <div className={styles.App}>
             {/* settings component there is just temporary */}
-            <Settings isCollaborative={true} />
             <p>This is the homepage</p>
-            <SockJsClient
-                url="http://localhost:8080/gs-guide-websocket"
-                topics={['/topic/hello']}
-                onMessage={(msg) => {
-                    console.log('this is the message', msg);
-                    onMessageRecieve(msg);
-                }}
-                ref={(client) => {
-                    setClientRef(client);
-                }}
-                onConnect={() => {
-                    console.log('successuflly connected to websocket');
-                    setClientConnected(true);
-                }}
-                onDisconnect={() => setClientConnected(false)}
-                debug={false}
-            />
-            <button onClick={() => sendMessage()}>send a message</button>
-            {messages.map((elem) => (
-                <div>{elem}</div>
-            ))}
+            <button onClick={navigateToCollaborativeSettings}>
+                Collaborative Game
+            </button>
+            <button onClick={navigateToCompetitiveSettings}>
+                Competitive Game
+            </button>
         </div>
     );
 });
