+++ frontend/src/index.tsx
@@ -5,6 +5,8 @@ import './index.css';
 import Homepage from './components/homepage/Homepage';
 import CompetitivePage from './components/competitive/Competitive';
 import CollaborativePage from './components/collaborative/Collaborative';
+import Settings from './components/settings/Settings';
+
 import reportWebVitals from './reportWebVitals';
 /* Remember to delete this */
 import Crossword from './components/shared/Crossword';
@@ -15,6 +17,12 @@ ReactDOM.render(
             <Route exact path="/" component={Homepage} />
             <Route exact path="/competitive" component={CompetitivePage} />
             <Route exact path="/collaborative" component={CollaborativePage} />
+            <Route exact path="/competitive-settings">
+                <Settings isCollaborative={false} />
+            </Route>
+            <Route exact path="/collaborative-settings">
+                <Settings isCollaborative={true} />
+            </Route>
         </HashRouter>
     </React.StrictMode>,
     document.getElementById('root')
